+++ test/adventofcode/twenty19/Day7Test.kt
@@ -21,7 +21,7 @@ class Day7Test {
      */
     @Test
     fun TestController1() {
-        val memory = arrayOf(3, 15, 3, 16, 1002, 16, 10, 16, 1, 16, 15, 15, 4, 15, 99, 0, 0)
+        val memory = arrayOf(3L, 15, 3, 16, 1002, 16, 10, 16, 1, 16, 15, 15, 4, 15, 99, 0, 0)
         val sequence = arrayOf(4, 3, 2, 1, 0)
 
         val computerA = IntcodeComputer()
@@ -46,7 +46,7 @@ class Day7Test {
      */
     @Test
     fun TestController2() {
-        val memory = arrayOf(3, 23, 3, 24, 1002, 24, 10, 24, 1002, 23, -1, 23, 101, 5, 23, 23, 1, 24, 23, 23, 4, 23, 99, 0, 0)
+        val memory = arrayOf(3L, 23, 3, 24, 1002, 24, 10, 24, 1002, 23, -1, 23, 101, 5, 23, 23, 1, 24, 23, 23, 4, 23, 99, 0, 0)
 
         val computerA = IntcodeComputer()
         val computerB = IntcodeComputer()
@@ -69,7 +69,7 @@ class Day7Test {
      */
     @Test
     fun TestController3() {
-        val memory = arrayOf(3, 31, 3, 32, 1002, 32, 10, 32, 1001, 31, -2, 31, 1007, 31, 0, 33, 1002, 33, 7, 33, 1, 33, 31, 31, 1, 32, 31, 31, 4, 31, 99, 0, 0, 0)
+        val memory = arrayOf(3L, 31, 3, 32, 1002, 32, 10, 32, 1001, 31, -2, 31, 1007, 31, 0, 33, 1002, 33, 7, 33, 1, 33, 31, 31, 1, 32, 31, 31, 4, 31, 99, 0, 0, 0)
 
         val computerA = IntcodeComputer()
         val computerB = IntcodeComputer()
@@ -96,7 +96,7 @@ class Day7Test {
      */
     @Test
     fun testControllerPart2_1() {
-        val memory = arrayOf(3, 26, 1001, 26, -4, 26, 3, 27, 1002, 27, 2, 27, 1, 27, 26, 27, 4, 27, 1001, 28, -1, 28, 1005, 28, 6, 99, 0, 0, 5)
+        val memory = arrayOf(3L, 26, 1001, 26, -4, 26, 3, 27, 1002, 27, 2, 27, 1, 27, 26, 27, 4, 27, 1001, 28, -1, 28, 1005, 28, 6, 99, 0, 0, 5)
 
         val computerA = IntcodeComputer()
         val memoryA = memory.copyOf()
@@ -109,13 +109,13 @@ class Day7Test {
         val computerE = IntcodeComputer()
         val memoryE = memory.copyOf()
 
-        var resultA: Int
-        var resultB: Int
-        var resultC: Int
-        var resultD: Int
-        var resultE = 0
+        var resultA: Long
+        var resultB: Long
+        var resultC: Long
+        var resultD: Long
+        var resultE = 0L
 
-        var finalOutput = 0
+        var finalOutput = 0L
 
         var loop = 0
         do {
@@ -134,12 +134,12 @@ class Day7Test {
                 resultE = computerE.executeToOutput(memoryE, arrayOf(resultD))
             }
 
-            if (resultE != -9999) {
+            if (resultE != -9999L) {
                 finalOutput = resultE
             }
 
             loop++
-        } while (resultE != -9999)
+        } while (resultE != -9999L)
 
         assertEquals(139629729, finalOutput)
     }
@@ -151,7 +151,7 @@ class Day7Test {
      */
     @Test
     fun testControllerPart2_2() {
-        val memory = arrayOf(3, 52, 1001, 52, -5, 52, 3, 53, 1, 52, 56, 54, 1007, 54, 5, 55, 1005, 55, 26, 1001, 54, -5, 54, 1105, 1, 12, 1, 53, 54, 53, 1008, 54, 0, 55, 1001, 55, 1, 55, 2, 53, 55, 53, 4, 53, 1001, 56, -1, 56, 1005, 56, 6, 99, 0, 0, 0, 0, 10)
+        val memory = arrayOf(3L, 52, 1001, 52, -5, 52, 3, 53, 1, 52, 56, 54, 1007, 54, 5, 55, 1005, 55, 26, 1001, 54, -5, 54, 1105, 1, 12, 1, 53, 54, 53, 1008, 54, 0, 55, 1001, 55, 1, 55, 2, 53, 55, 53, 4, 53, 1001, 56, -1, 56, 1005, 56, 6, 99, 0, 0, 0, 0, 10)
 
         val computerA = IntcodeComputer()
         val memoryA = memory.copyOf()
@@ -164,13 +164,13 @@ class Day7Test {
         val computerE = IntcodeComputer()
         val memoryE = memory.copyOf()
 
-        var resultA: Int
-        var resultB: Int
-        var resultC: Int
-        var resultD: Int
-        var resultE = 0
+        var resultA: Long
+        var resultB: Long
+        var resultC: Long
+        var resultD: Long
+        var resultE = 0L
 
-        var finalOutput = 0
+        var finalOutput = 0L
 
         var loop = 0
         do {
@@ -189,12 +189,12 @@ class Day7Test {
                 resultE = computerE.executeToOutput(memoryE, arrayOf(resultD))
             }
 
-            if (resultE != -9999) {
+            if (resultE != -9999L) {
                 finalOutput = resultE
             }
 
             loop++
-        } while (resultE != -9999)
+        } while (resultE != -9999L)
 
         assertEquals(18216, finalOutput)
     }
